Begin of Major changes to architecture.  Viewmodels moveing to ECSLib.  Cargo code Reworked.   currently not building.

+++ Pulsar4X/Pulsar4X.ECSLib/StaticData/CargoTypeSD.cs
@@ -15,7 +15,7 @@ namespace Pulsar4X.ECSLib
         Guid ID { get; }
         string Name { get; }
         Guid CargoTypeID { get;  }
-        float Mass { get;  }
+        int Mass { get;  }
     }
 }
